Move generate_wins_draw_losses to bt_output (it's an output function, not a calculation)
