+++ freqtrade/optimize/optimize_reports/optimize_reports.py
@@ -57,16 +57,6 @@ def generate_rejected_signals(preprocessed_df: Dict[str, DataFrame],
     return rejected_candles_only
 
 
-def generate_wins_draws_losses(wins, draws, losses):
-    if wins > 0 and losses == 0:
-        wl_ratio = '100'
-    elif wins == 0:
-        wl_ratio = '0'
-    else:
-        wl_ratio = f'{100.0 / (wins + draws + losses) * wins:.1f}' if losses > 0 else '100'
-    return f'{wins:>4}  {draws:>4}  {losses:>4}  {wl_ratio:>4}'
-
-
 def _generate_result_line(result: DataFrame, starting_balance: int, first_column: str) -> Dict:
     """
     Generate one result dict, with "first_column" as key.
